+++ src/app/_components/Navigation/Navigation.tsx
@@ -1,56 +1,28 @@
 'use client';
 
-import { getMyPlans } from '@/apis/client/getMyPlans';
 import { Icon } from '@/components';
-import { ajajaToast } from '@/components/Toaster/customToast';
-import { maxPlan } from '@/constants/plan';
-import { canMakeNewPlanStore } from '@/stores/canMakeNewPlanStore';
-import { isMyPlanStore } from '@/stores/isMyPlanStore';
+import { address } from '@/constants';
 import { checkIsSeason } from '@/utils/checkIsSeason';
-import { checkThisYear } from '@/utils/checkThisYear';
 import classNames from 'classnames';
-import { hasCookie } from 'cookies-next';
 import Link from 'next/link';
-import { usePathname } from 'next/navigation';
-import { useEffect, useState } from 'react';
-import { useRecoilState } from 'recoil';
+import { useNavigation } from '../hooks';
 import './index.scss';
 
 export default function Navigation({ hasAuth }: { hasAuth: boolean }) {
-  const pathName = usePathname();
-  const [isLogin, setIsLogin] = useState(hasAuth);
-  const [canMakeNewPlan, setCanMakeNewPlan] =
-    useRecoilState(canMakeNewPlanStore);
-  const [isMyPlan] = useRecoilState(isMyPlanStore);
-  const isEdit = /^\/plans\/edit\/\d+/;
-  const isPlan = /^\/plans\/\d+/;
-  const isRemind = /^\/reminds\/.*$/;
-  const isFeedback = /^\/feedback\/\d+/;
-  const isFeedbackEvaluate = /^\/feedback\/evaluate/;
+  const {
+    pathName,
+    canMakeNewPlan,
+    isLogin,
+    isMyPlan,
+    isRealLogin,
+    handleCreate,
+  } = useNavigation({
+    hasAuth,
+  });
+  const { isEdit, isFeedback, isFeedbackEvaluate, isPlan, isRemind } = address;
 
-  if (!hasCookie('auth')) {
-    setTimeout(() => {
-      setIsLogin(hasCookie('auth'));
-    }, 1000);
-  }
+  isRealLogin();
 
-  const handleCreate = () => {
-    if (!canMakeNewPlan) {
-      ajajaToast.error('생성할 수 있는 계획의 수가 최대입니다.');
-    } else if (!checkIsSeason()) {
-      ajajaToast.error('계획을 작성할 수 있는 시즌이 아닙니다.');
-    }
-  };
-
-  useEffect(() => {
-    async function isMaxPlan() {
-      const data = await getMyPlans();
-      if (data.data[0]?.year === checkThisYear()) {
-        setCanMakeNewPlan(!!(maxPlan - data.data[0].getPlanList.length));
-      }
-    }
-    isMaxPlan();
-  }, [setCanMakeNewPlan]);
   return (
     <div className={classNames('navigation')}>
       <Link href="/home" className={classNames('navigation-icon')}>
